requirements.txt
asgiref==3.5.2
autopep8==1.6.0
bcrypt==3.2.2
black==22.6.0
boto3==1.24.42
botocore==1.27.42
cffi==1.15.1
click==8.1.3
colorama==0.4.5
Django==4.0.6
django-cors-headers==3.13.0
django-storages==1.12.3
djangorestframework==3.13.1
jmespath==1.0.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.9.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.11
